requirements.txt
# Core ML libraries
torch
torchvision
numpy
pandas
scikit-learn
matplotlib

# Jupyter for experimentation
jupyter

# Backend
django
djangorestframework
pillow   # Required for handling image uploads in Django
djangorestframework-simplejwt # Added for token-based authentication
django-cors-headers # Added for CORS configuration
django-filter # Added for filtering


# Optional utilities (safe to include)
tqdm
python-multipart

# Frontend note (install via npm/yarn, not pip):
# Run `npm install` inside src/frontend to install:
# @emotion/react
# @emotion/styled
# @mui/icons-material
# @mui/material
# axios
# jwt-decode
# react
# react-dom
# react-router-dom
# (plus dev deps: @vitejs/plugin-react, typescript, eslint, vite, etc.)